Abuse @Published to stop webviews reloading
There is an issue around tying the WebView state to the UI state, via a bound URL.
When a view update happens that bound url is reloaded, which is fine unless
the webview is showing another URL. In that case the bound url is reloaded and the user is kicked back

What seems to be working is removing the binding. The url is still marked as @Published so I can
mutate it inside a struct, but it's not marked with Binding so isn't tying the WebView into SwiftUI view updates.
This should mean the incorrect URLs get loaded as they aren't know until runtime, but SwiftUI is re-creating all the
webviews (multiple times) when there is a view update, probably because aside from the URL i'm updating some bools
that _are_ participating in view updates.

The one thing that makes me think this is kind of ok is SwiftUIs default behaviour is to recreate views all the time, but
using @Published without an associated binding seems incorrect.

This might break in the future, hopefully at compile time and not runtime in some OS update