WIP, webview refactor
Have the WKWebView in a Model object managed by the containing view. This model and WKWebView
live for the lifetime of the View object. A very simple WebView just draws this WKWebView
held in the model.
There are no bindings at play, to load a page a function call is used.
Next is figuring out how to put this thing in a container with a URL _binding_